requirements.txt
anyio==3.6.2
asgiref==3.6.0
bidict==0.22.1
click==8.1.3
colorama==0.4.6
Django==4.1.7
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
sniffio==1.3.0
sqlparse==0.4.3
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
